Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.27
transformers==4.38.1
mecab-python3==1.0.8
fugashi==1.3.0
ipadic==1.0.0
unidic-lite==1.0.8
torch==2.2.1
numpy==1.26.3
scikit-learn==1.6.1
jaconv==0.4.0
pykakasi==2.3.0